+++ box.config.js
@@ -23,6 +23,9 @@ module.exports = function (config) {
     output: 'dist',
     publicPath: '/common/',
     port: 8888,
+    env: {
+      MY_ENV: 'my-env'
+    },
     alias: {
       '@': resolve('src'),
       '@src': resolve('src')