+++ box.config.js
@@ -10,14 +10,15 @@ module.exports = function (config) {
    * @param {function} chainWebpack
    * @param {string} entry 入口
    * @param {string} output 出口
-   *
    * @param {string} publicPath
    * @param {string} port 端口
    * @param {object} eslint eslint 配置
    * @param {object} stylelint stylelint 配置
    * @param {object} eslint eslint 配置
    * @param {object} alias 配置别名
-   * @param {Boolean} filenameHashing 文件名是否使用 hash
+   * @param {object} env 配置通用变量，可以在 node 跟 web 之间共同使用
+   * @param {Boolean} filenameHashing 文件名是否使用 hash，当文件发生变动的时候 filename 才会改变
+   * @param {Boolean} css 配置 css
    */
   return {
     entry: 'src/main.js',
@@ -30,6 +31,34 @@ module.exports = function (config) {
     alias: {
       '@': resolve('src'),
       '@src': resolve('src')
+    },
+    resources: {
+      less: {
+        patterns: [
+          path.resolve(__dirname, './src/global/*.less')
+        ]
+      },
+      scss: {
+        patterns: [
+          path.resolve(__dirname, './src/global/*.scss')
+        ]
+      }
+    },
+    css: {
+      sourceMap: true,
+      loaderOptions: {
+        css: {},
+        less: {
+          globalVars: {
+            gray: '#ccc'
+          }
+        },
+        sass: {},
+        postcss: {},
+        stylus: {}
+      },
+      isCssModule: false, // 是否对css进行模块化处理
+      needInlineMinification: false // 是否需要压缩css
     },
     filenameHashing: true,
     eslint: {