+++ config/ProgressBarPlugin.js
@@ -2,8 +2,12 @@
 
 module.exports = ({ config, resolve, options }) => {
   const ProgressBarPlugin = require('progress-bar-webpack-plugin')
+  const chalk = require('chalk')
   return () => {
     config.plugin('ProgressBarPlugin')
-      .use(ProgressBarPlugin)
+      .use(ProgressBarPlugin, [{
+        format: '  build [:bar] ' + chalk.green.bold(':percent') + ' (:elapsed seconds)',
+        clear: false
+      }])
   }
 }