+++ config/PreloadWebpackPlugin.js
@@ -3,21 +3,20 @@
 module.exports = ({ config, resolve, options }) => {
   const PreloadWebpackPlugin = require('preload-webpack-plugin')
   return () => {
-    config.plugin('preload')
+    config
+      .plugin('preload')
       .use(PreloadWebpackPlugin, [{
         rel: 'preload',
-        as(entry) {
-          if (/\.css$/.test(entry)) return 'style'
-          if (/\.woff$/.test(entry)) return 'font'
-          if (/\.png$/.test(entry)) return 'image'
-          return 'script'
-        }
+        include: 'initial',
+        fileBlacklist: [/\.map$/, /hot-update\.js$/]
       }])
+      .after('html')
     config
       .plugin('prefetch')
       .use(PreloadWebpackPlugin, [{
         rel: 'prefetch',
         include: 'asyncChunks'
       }])
+      .after('html')
   }
 }